context-graph-mvp/requirements.txt
neo4j>=5.0.0
langchain>=0.1.0
langchain-community>=0.0.1
chromadb>=0.4.0
anthropic>=0.30.0
google-genai>=1.0.0
yfinance>=0.2.0
pdfplumber>=0.10.0
streamlit>=1.30.0
sentence-transformers>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
pyvis>=0.3.0
fredapi>=0.5.0
pytest>=7.0.0
